Collapse AzureAIChatToolJson into ChatClient
